apply DBitV to incompressible sequences

+++ include/am/succinct/fm-index/wavelet_tree.hpp
@@ -19,7 +19,7 @@ class WaveletTree
 public:
     typedef CharT char_type;
 
-    WaveletTree(size_t alphabet_num, bool support_select, bool dense)
+    WaveletTree(uint64_t alphabet_num, bool support_select, bool dense)
         : alphabet_num_(alphabet_num)
         , alphabet_bit_num_()
         , support_select_(support_select)
@@ -62,7 +62,7 @@ public:
     virtual size_t length() const = 0;
     virtual size_t allocSize() const = 0;
 
-    inline size_t getAlphabetNum() const
+    inline uint64_t getAlphabetNum() const
     {
         return alphabet_num_;
     }
@@ -96,19 +96,19 @@ public:
         istr.read((char *)&dense_, sizeof(dense_));
     }
 
-    static size_t getAlphabetNum(const char_type *char_seq, size_t len)
+    static uint64_t getAlphabetNum(const char_type *char_seq, size_t len)
     {
-        size_t num = 0;
+        uint64_t num = 0;
         for (size_t i = 0; i < len; ++i)
         {
-            num = std::max((size_t)char_seq[i] + 1, num);
+            num = std::max((uint64_t)char_seq[i] + 1, num);
         }
 
         return num;
     }
 
 protected:
-    size_t alphabet_num_;
+    uint64_t alphabet_num_;
     size_t alphabet_bit_num_;
     bool support_select_;
     bool dense_;
